+++ tortoise/fields.py
@@ -58,6 +58,10 @@ class Field:
             return value
         return self.type(value)
 
+    @property
+    def required(self):
+        return self.default is None and not self.null and not self.generated
+
 
 class IntField(Field):
     """